requirements.txt
werkzeug
pymongo
flask
flask_restful
flask_cors
flask_jwt_simple
waitress
fuzzywuzzy
pandas
tinydb
xlsxwriter
xlrd
tqdm
mysql-connector-python
requests